#117 HistoryLogService -> HistoryLogService, HistoryLogServiceImpl
